Separate util classes to own sub-packages.
