+++ src/main/kotlin/hu/netsurf/erp/common/util/sanitization/InputSanitizer.kt
@@ -1,4 +1,4 @@
-package hu.netsurf.erp.common.util
+package hu.netsurf.erp.common.util.sanitization
 
 import org.springframework.stereotype.Component
